python_server/requirements_llm.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asgi-csrf==0.11
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-default-group==1.2.4
comm==0.2.2
datasette==0.65
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
flexcache==0.3
flexparser==0.4
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hupper==1.12.1
idna==3.10
ijson==3.3.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
janus==1.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llm==0.18
llm-claude-3==0.9
llm-gemini==0.4.2
llm-groq==0.5
llm-together==0.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mergedeep==1.3.4
nest-asyncio==1.6.0
openai==1.55.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-ulid==3.0.0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.37
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
together==0.2.9
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
